third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-overflow-cssomview-expected.txt

This is a testharness.js-based test.
FAIL Test cssom-view API for FIELDSET assert_equals: expected 400 but got 418
Harness: the test ran to completion.
